requirements.txt
django
pytest-django
pytest-factoryboy
Faker
mixer